requirements.txt
numpy
scikit-learn
tensorflow-gpu==1.13.1